refactor(direct-control): extract setup lifecycle types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only single-player setup input, setup option value, prepared-setup result, prepared-start input, single-player start result, and single-player run input/result type ownership from packages/civ7-direct-control/src/index.ts into the concrete setup owners: packages/civ7-direct-control/src/setup/prepare.ts, packages/civ7-direct-control/src/setup/start.ts, and packages/civ7-direct-control/src/setup/run.ts. Keeps facade type re-exports in index.ts.

Preserves public type contracts and does not change setup preparation/start/run source strings, approval behavior, readback verification, setup lifecycle mutation behavior, runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Tactical, operation, ready, public procedure schema, and telemetry type ownership remain pending separate owner slices.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- setup-and-lifecycle.test.ts public-api.test.ts restart-lifecycle.test.ts
- rg -n "setup|single-player|startPrepared|prepareCiv7" packages/cli/test/commands packages/cli/src/commands
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, setup lifecycle behavior, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
