+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -419,6 +419,13 @@ runtime/direct-control claims.
         hotseat runtime proof, AI ingestion, CLI semantic projection,
         Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance
         pending.
+  - [x] 4.8.17 Extract setup lifecycle type owners while keeping public facade
+        type re-exports in `index.ts`, preserving prepare/start/run
+        input/result public contracts while leaving tactical, operation, and
+        ready result types, procedure schemas, operation/proof telemetry,
+        hotseat runtime proof, AI ingestion, CLI semantic projection,
+        Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance
+        pending.
 - [ ] 4.9 Import or explicitly cite oRPC architecture authority before any
       oRPC implementation; current support branch has no tracked
       `.agents/skills/civ7-orpc-control-architecture` or
